Environment Space Projector
this is the project files of a addon that i put hard work into for a sandbox game that i enjoy! it basically projects stars, planets, nebulas in a map of your game up in the air! you should read more about it at https://steamcommunity.com/sharedfiles/filedetails/stats/3449439032 i think you will like it!